renamed scide decompression file
